Create general-use panel macro with a toggling visual caret
Generalize the panel that currently is only being used for event
participants, to remove event-specific stuff and also add a
visual caret to the header to indicate more clearly that the
panel can be collapsed.

The tooltip text is added in JS because it doesn't have any
meaning for non-JS users.

Bug: https://phabricator.wikimedia.org/T200551