Fix countable error for PHP 7.2
Switch to normal PHPUnit instead of the Symfony bridge, for
testing in CI. We'll return to the bridge after upgrading to
Symfony 4.

Doctrine\ORM\EntityRepository::findOneBy() returns NULL if nothing
is found, and as of PHP 7.2 this isn't countable.

Add PHP 7.2 to CI, for future bug-catching, and upgrade phpdoc.

Silence Symfony deprecation warnings; we'll re-enable after upgrading.