API endpoint to get all wikis adhering to EventWiki::VALID_WIKI_PATTERN
This will be used in clientside JavaScript for autocompletion on wiki
inputs. We can't go by the Sitematrix API because newly introduced wikis
may be production but not yet on the replicas, causing Grant Metrics to
error out.

This commit also adds ext-json as a dependency in composer.json, along
with some unrelated minor code aesthetics.