Rework setupAddRemove() in preparation for category support
This makes grantmetrics.application.setupAddRemove() manipulate the name
and id attributes when cloning the template, rather than rebuilding it.
This is because the category form has nested associations so we can't
build the name attributes in the same way as organizers and wikis.

Don't make organizer inputs required (blank ones are automatically
removed), and disable autocompletion that conflicts with Typeahead.

gitignore development assets.

Remove 'use' statements of unused classes in DefaultController.