Use CSS to show invalid input icon; autoscroll to invalid form
With category support, we have multiple inputs per row in the view.
It is easier and cleaner to use CSS to show the invalid input icon,
rather than adding conditionals for each individual input.

This commit also removes some unnecessary code for checking if a field
is invalid, and moves any disabling of autocompletion to the form tag.