Use IDs in routes instead of titles; disallow IDs in titles
Because we're disallowing IDs in titles (and there are no existing
events/programs with numerical titles), we can test if the title is a
string versus a number to make the routing backwards-compatible. This
is only done for routes likely to be shared, such as the event page.

As a result of the new routing, we don't need to validate program/event
titles do not contain 'edit', 'new', etc., so this logic and the
associated error message have been removed.

Bug: https://phabricator.wikimedia.org/T192748