Add pageviews stats
Add two new stats for pageviews of pages created, and average
daily pageviews (over up to the last 30 days) of pages edited.
Neither are shown anywhere yet.

A new repository is added for this, with methods for each of
the pageviews methods of the API. Only the required parameters
are exposed.

Bug: https://phabricator.wikimedia.org/T206817