Prevent empty page IDs when decompressing bzcompressed data
When we switched to two page ID columns the 2nd one is initally
populated with NULLs. When this is retrieved and decompressed
it results in an empty string, which was being passed through
in the pageIds arrays and breaking SQL.

This patch fixes it by removing any empty elements in the returned
array.

Bug: https://phabricator.wikimedia.org/T213230