Store the status of Jobs and report failures
Change the job_started (bool) column into a tinyint that represents the
status of the Job. We do this so that we can flag it as failed.

This is basically a regression fix. Before we deployed the 'proper' job
queue system, a failed job would be reported in the repose of the
request to start the job. Now it is a true background process so we must
store the job state in order to determine if it failed.