Fix phpcs test within Travis
The Travis build is put in a directory called 'build'. Our repo also has
a directory called 'build' that we want to exclude. Relative paths
apparently also don't work, so instead we specify exactly what needs to
be excluded within the build directory.

Also removing the public_html symlink, a relic from when we were still
on Toolforge.