Fix datestamps in queries and display on Edit List page
Rename Event::getStartWithTimezone() (and related methods) to
getStartUTC(), to avoid confusion. This commit changes all queries to
use the UTC variant, as it should. In hindsight, we should have stored
datestamps in UTC and only converted for display purposes, but it's
perhaps too late for that.

Make sure the Edit List page shows timestamps in the event's timezone
and not UTC.

Tests were updated to account for the longstanding bugs of using the
non-UTC timestamps in SQL queries.

Other minor code cleanup.

Bug: T215926