Improve logging in EventProcessor, using Stopwatch to log runtimes
When events are processed via the UI ('Update Data' button), the output
is logged var/logs/event_processor.log. When running an update directly
via 'php bin/console app:process-event', colourized output is sent to
stdout.

This doesn't include actual query runtimes, as requested by T205322, but
it does reveal which pieces of the update process are the slowest.

Bug: T205322