Store files uploaded as their own type of page IDs, update queries
This simplifies the code and makes the queries faster. Additionally, the
revision table retains the original uploader of the file, while going
off of the image table we only get the most recent version (which may be
by a non-participant).

On the backend, the "type" of page ID is called "pagesFiles" -- the name
chosen so it can use the helpers for "pagesCreated" and "pagesImproved"
that reference class property by the given 'type'.

Only use revision_userindex when filtering by user.