Display an alert before navigating if the form is 'dirty'
Make sure that if any form is in a dirty state, or if there
are any errors on a form, we ask the user whether they're
certain they want to navigate away.

Make sure that this does not prevent submission of the form
and that we do not interrupt regular navigation.

Bug: https://phabricator.wikimedia.org/T212804